feat(back): calculating delta

+++ back/src/Liane/Liane.Api/Object/Annotation.cs
@@ -6,7 +6,7 @@ namespace Liane.Api.Object
     // TODO: comment explaining Annotation object
     public sealed class Annotation
     {
-        public Annotation(ImmutableList<int> distance, ImmutableList<int> duration, ImmutableList<int> weight, ImmutableList<int> nodes, ImmutableList<float>? speed, ImmutableList<int> datasources, ImmutableList<string> datasourceNames)
+        public Annotation(ImmutableList<long> distance, ImmutableList<long> duration, ImmutableList<long> weight, ImmutableList<long> nodes, ImmutableList<float>? speed, ImmutableList<long> datasources, ImmutableList<string> datasourceNames)
         {
             Distance = distance;
             Duration = duration;
@@ -18,20 +18,20 @@ namespace Liane.Api.Object
         }
 
         // The distance, in metres, between each pair of coordinates.
-        public ImmutableList<int> Distance { get; }
+        public ImmutableList<long> Distance { get; }
         // The duration between each pair of coordinates, in seconds. Does not include the duration of any turns.
-        public ImmutableList<int> Duration { get; }
+        public ImmutableList<long> Duration { get; }
         // The weights between each pair of coordinates. Does not include any turn costs. 
-        public ImmutableList<int> Weight { get; }
+        public ImmutableList<long> Weight { get; }
 
         // The OSM node ID for each coordinate along the route, excluding the first/last user-supplied coordinates
-        public ImmutableList<int> Nodes { get; }
+        public ImmutableList<long> Nodes { get; }
         // Convenience field, calculation of distance / duration rounded to one decimal place
         public ImmutableList<float> ?Speed { get; }
         
         // The index of the datasource for the speed between each pair of coordinates. 0 is the default profile, other values are supplied via --segment-speed-file to osrm-contract or osrm-customize.
         // String-like names are in the metadata.datasource_names array.
-        public ImmutableList<int> Datasources { get; }
+        public ImmutableList<long> Datasources { get; }
         //  The names of the datasources used for the speed between each pair of coordinates. lua profile is the default profile, other values arethe filenames supplied via --segment-speed-file to osrm-contract or osrm-customize
         public ImmutableList<string> DatasourceNames { get; } // ATTENTION contenu dans metadata
